refactor: create util function for fetch options
